requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
langchain==0.1.20
langchain-google-genai
faiss-cpu==1.12.0
newsapi-python
sentence-transformers